Exit from with the function to isolate against edits
Unixen shells read SCRIPTS one line at a time (except fish), which means
that editing a file while it is being exectuted elsewhere will almost
certainly affect the running instance.  While that's not usually
a problem, it is with any long running process.

However unixen shells read in FUNCTIONS entirely at their declaration,
so to avoid having changes to a script affect other running instances,
move everything into functions and call the main function as the last
thing in a script.

IT IS STILL NECESSARY to call `exit` from within the FUNCTION to keep
the shell from returning to the possibly changed script.