Refactor (simplify,clarify); embetter output
Save the declaration and all the keys and values to $@ so we can use
variables without overwriting them before we get their information.
But we need a couple of variables to do that, so special case handling
those variables and only use the generalized handling otherwise.

Wrap very long variable descriptions.