Use `$(</dev/stdin)` instead of `$(cat)`
`cat` is an external executable which must be fork+exec'd. The `$(<…)`
facility is entirely internal and with the OS provided `/dev/stdin` this
uses fewer resources without compromising readability very much and with
the same practical result.

Of course patching ksh to provide a HEREDOC-as-string facility would
probably be better.