wrap doas calls with as-root
`doas` gives only one (1) opportunity to enter a password. `as-root`
gives three (3), BUT makes sure that a password is needed before trying to
get one.

Several of the scripts changed made more rudimentary attempts to do the
same thing but didn't check to see if one was necessary.

This moves all of that to a single place which also makes it easier to
change in the future if a different executable is used (sudo?).