Have better screen redrawing
[debug] use correct function names for un/setting alternate screen

Change function name from set-buffer (which is what it originally was
going to do before I realized the drawing was fast enough not to need
buffering) to update-screen (which is what it does).

Move update-screen out of signal handling, though it shouldn't really
matter since it's KSH and the function isn't really run in the interrupt
handler, but it does seem to fix some minor glitches.