Break out machine specific setup (functions,scripts,etc)
In order to share dotfiles across machines, we must either provide the
same environment on every machine or separate out the bits that don't
make sense on some and put those in machine specific places.

I've been doing the first, but it doesn't really make sense on a
headless system to provide a music system or such. So now I'm trying the
second.

Basically, the specific bits will be moved to $K/$HOST/??? and sucked in
from there. In some cases it MIGHT be useful to also have a $K/share and
link to those when specific bits are shared across systems but not
universally.